requirements.txt
requests
python-dotenv
pydantic
groq